Remove ContributionViewModel
